It compiles! Store the future and poll it instead of creating multiple new ones
