+++ src/stream/mod.rs
@@ -303,6 +303,7 @@
 pub use empty::{empty, Empty};
 pub use from_fn::{from_fn, FromFn};
 pub use from_iter::{from_iter, FromIter};
+pub use successor::{successor, Successor};
 pub use once::{once, Once};
 pub use repeat::{repeat, Repeat};
 pub use repeat_with::{repeat_with, RepeatWith};
@@ -316,6 +317,7 @@ mod from_iter;
 mod once;
 mod repeat;
 mod repeat_with;
+mod successor;
 
 cfg_unstable! {
     mod double_ended_stream;
