+++ src/stream/stream/mod.rs
@@ -57,7 +57,6 @@ mod partial_cmp;
 mod position;
 mod scan;
 mod skip;
-mod successor;
 mod skip_while;
 mod step_by;
 mod take;
